Create & extend abstract for username/password authentication methods
